requirements.txt
# Embeddings (pinned for torch 2.2.x / Python 3.10 compatibility)
sentence-transformers==2.7.0
transformers==4.44.2
numpy<2

# Vector DB
chromadb

# Keyword search
rank-bm25

# Document parsing
pymupdf

# API layer
fastapi
uvicorn[standard]

# Frontend
streamlit

# Config & utilities
pydantic-settings
pyyaml
httpx
python-multipart
